src/components/data_ingestion/requirements.txt
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.1
fastparquet==2024.5.0
pyyaml==6.0.1
